% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords2bbox.R
\name{coords2bbox}
\alias{coords2bbox}
\title{Grabs lat and lon and returns bbox}
\usage{
coords2bbox(data, longitude = "Longitude", latitude = "Latitude")
}
\arguments{
\item{data}{A `data.frame` with coordinates in it.}

\item{longitude}{The name of the column for the x-axis coordinate}

\item{latitude}{The name of the column for the y-axis coordinate}
}
\value{
A bounding box (left, bottom, right, top)
}
\description{
Quick function to return a bounding box from a `data.frame` that
 has longitude and latitude columns.
}
\author{
Rodney Dyer <rjdyer@vcu.edu>
}
