% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_centroids.R
\name{partition_centroids}
\alias{partition_centroids}
\title{Partition coordinates by groups and find centers}
\usage{
partition_centroids(data, stratum = "Population",
  latitude = "Latitude", longitude = "Longitude")
}
\arguments{
\item{data}{A data frame object with coordinates and a column to be used as
stratum designations.}

\item{stratum}{The name of the column to use as the partition.}

\item{latitude}{The name of the column with the Y-coordinate (default = Latitude)}

\item{longitude}{The name of the column with the X-coordinate (default = Longitude)}
}
\value{
A data.frame object with strata and latitude and longitude
}
\description{
This function partitions the data.frame or tibble into subgroups
  and returns the names and centroids of the strata.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
