% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_centroids.R
\name{strata_centroids}
\alias{strata_centroids}
\title{Partition coordinates by strata and find mean coordinates}
\usage{
strata_centroids(data, stratum = "Population", longitude = "Longitude",
  latitude = "Latitude", mode = c("data.frame", "SpatialPoints",
  "sf")[1])
}
\arguments{
\item{data}{A data frame object with coordinates and a column to be used as
stratum designations.}

\item{stratum}{The name of the column to use as the partition.}

\item{longitude}{The name of the column with the X-coordinate (default = Longitude)}

\item{latitude}{The name of the column with the Y-coordinate (default = Latitude)}
}
\value{
A data.frame object with strata and latitude and longitude
}
\description{
This function partitions the data.frame or tibble into subgroups
  and returns the names and centroids of the strata.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
